+++ tinyredis/redis.d
@@ -107,7 +107,7 @@ private :
     void receive(Socket conn, ref byte[] buffer)
     {
         byte[1024 * 16] buff;
-        long len = conn.receive(buff);
+        size_t len = conn.receive(buff);
         
         if(len == 0)
             throw new ConnectionException("Server closed the connection!");
